+++ backend/src/utils/enums/violations.enum.ts
@@ -2,5 +2,6 @@ export enum Violation {
     CLIENTSFLAG = 'clients-flag',
     USERSFLAG = 'users-flag',
     USERSSTATUS = 'users-status',
-    TICKETSFLAG = 'tickets-flag'
+    TICKETSFLAG = 'tickets-flag',
+    MAINTENANCE_TRAFFIC = 'maintenance-traffic',
 }