+++ backend/tests/unit-tests/validation/common.validation.test.ts
@@ -1,5 +1,5 @@
 import { MaintenanceMode } from "../../../src/utils/enums/maintenance-mode.enum";
-import * as CustomValidators from "../../../src/utils/customValidator.utils";
+import * as CommonValidators from "../../../src/validation/common.validation";
 import { EnvMode } from "../../../src/utils/enums/env-mode.enum";
 import { MailSource } from "../../../src/utils/enums/mail-source.enum";
 import { ApiKeyStatus } from "../../../src/utils/enums/api-key-status.enum";
@@ -8,7 +8,7 @@ import { Flag } from "../../../src/utils/enums/flag.enum";
 import { CommonExceptionMessage } from "../../../src/utils/enums/common-exception-messages.enum";
 import { DemoMode } from "../../../src/utils/enums/demo-mode.enum";
 
-describe('CustomValidator tests, priority: no model specification', () => {
+describe('CommonValidators tests, priority: no model specification', () => {
 
     describe('Testing valid fn calls', () => {
 
@@ -18,7 +18,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -32,7 +32,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -46,7 +46,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -60,7 +60,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -74,7 +74,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -88,7 +88,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -102,7 +102,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = true;
 
             Object.values(mockParam_enumObj).forEach((value) => {
-                expect(CustomValidators.validateEnum(
+                expect(CommonValidators.validateEnum(
                     value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -112,7 +112,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
 
         test('fn: validatePathParam, params <param> = "valid-test-id"', () => {
             const mockParam_param = 'valid-test-id';
-            const testFn = CustomValidators.validatePathParam(mockParam_param);
+            const testFn = CommonValidators.validatePathParam(mockParam_param);
             const expectResult = true;
 
             expect(testFn).toBe(expectResult);
@@ -120,7 +120,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
 
         test('fn: validateTimestamp, params <timestamp> = "2025-01-01T14:00:00.000Z"', () => {
             const mockParam_timestamp = "2025-01-01T14:00:00.000Z";
-            const testFn = CustomValidators.validateTimestamp(mockParam_timestamp)
+            const testFn = CommonValidators.validateTimestamp(mockParam_timestamp)
             const expectResult = true;
 
             expect(testFn).toBe(expectResult);
@@ -128,7 +128,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
 
         test('fn: validateTimestampFilter, params <timestamps> = ["older", "younger"]', () => {
             const mockParam_timestamps = ['2025-01-01T14:00:00.000Z', '2025-01-10T13:00:00.000Z'];
-            const testFn = CustomValidators.validateTimestampFilter(mockParam_timestamps);
+            const testFn = CommonValidators.validateTimestampFilter(mockParam_timestamps);
             const expectResult = true;
 
             expect(testFn).toBe(expectResult);
@@ -144,7 +144,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = `support-invalid-entry#${mockParam_enumName}`;
 
             expect(() => {
-                CustomValidators.validateEnum(
+                CommonValidators.validateEnum(
                     mockParam_value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -153,13 +153,13 @@ describe('CustomValidator tests, priority: no model specification', () => {
         })
 
         test('fn: validateEnum, params: invalid <language> by key', () => {
-            const mockParam_value = { INVALID: MaintenanceMode.D013 };
+            const mockParam_value = { INVALID: MaintenanceMode.T011 };
             const mockParam_enumObj = MaintenanceMode;
             const mockParam_enumName = 'maintenanceMode';
             const expectResult = `support-invalid-entry#${mockParam_enumName}`;
             
             expect(() => {
-                CustomValidators.validateEnum(
+                CommonValidators.validateEnum(
                     mockParam_value,
                     mockParam_enumObj,
                     mockParam_enumName
@@ -172,7 +172,7 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const expectResult = CommonExceptionMessage.REQUIRED;
 
             expect(() => {
-                CustomValidators.validatePathParam(mockParam_arg);
+                CommonValidators.validatePathParam(mockParam_arg);
             }).toThrow(expectResult);
         })
 
@@ -180,21 +180,21 @@ describe('CustomValidator tests, priority: no model specification', () => {
             const mockParam_timestamp = "abc";
             const expectResult = 'support-invalid-entry#timestamp';
 
-            expect(() => { CustomValidators.validateTimestamp(mockParam_timestamp) }).toThrow(expectResult);
+            expect(() => { CommonValidators.validateTimestamp(mockParam_timestamp) }).toThrow(expectResult);
         })
 
         test('fn: validateTimestampFilter, params <timestamps> = ["younger", "older"]', () => {
             const mockParam_timestamps = ['2025-01-10T13:00:00.000Z', '2025-01-01T14:00:00.000Z'];
             const expectResult = 'support-invalid-entry#timestamps';
 
-            expect(() => { CustomValidators.validateTimestampFilter(mockParam_timestamps)}).toThrow(expectResult);
+            expect(() => { CommonValidators.validateTimestampFilter(mockParam_timestamps)}).toThrow(expectResult);
         })
 
         test('fn: validateTimestampFilter, params <timestamps> = ["single_value"]', () => {
             const mockParam_timestamps = ['2025-01-01T14:00:00.000Z'];
             const expectResult = 'support-invalid-entry#timestamps';
 
-            expect(() => { CustomValidators.validateTimestampFilter(mockParam_timestamps)}).toThrow(expectResult);
+            expect(() => { CommonValidators.validateTimestampFilter(mockParam_timestamps)}).toThrow(expectResult);
         })
     })
 })